requirements.txt
fastapi
uvicorn
pymupdf
python-docx
pandas
openpyxl
python-multipart
chromadb
flashrank
python-dotenv
google-genai